+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Release 4.0.0
+
+### New features
+- Supports FlowConfig to link digital IOs
+- Provide version of module via 'OnNewStatusModuleVersion'
+- Function 'getParameters' to provide PersistentData parameters
+- Check if features of module can be used on device and provide this via 'OnNewStatusModuleIsActive' event / 'getStatusModuleActive' function
+- Added functions 'removeForwardEventViaUI' and 'removeTriggerEventViaUI'. Furthermore 'removeTriggerEvent' and 'removeForwardEvent' require now the eventname as first parameter
+- Function to 'resetModule' to default setup
+
+### Improvements
+- New UI design available (e.g. selectable via CSK_Module_PersistentData v4.1.0 or higher), see 'OnNewStatusCSKStyle'
+- 'loadParameters' returns its success
+- 'sendParameters' can control if sent data should be saved directly by CSK_Module_PersistentData
+- Added UI icon
+- Changed log level of some messages from 'info' to 'fine'
+- Removed 'setForwardEvent' as it was not used
+
 ## Release 3.9.0
 
 ### New features